abstract request validator
